+++ react/SKUSelector.js
@@ -1,0 +1,3 @@
+import SKUSelector from './components/SKUSelector/index'
+
+export default SKUSelector