+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial version of SKU selector component.
 
 ### Fixed
+- Fix bug on Slider when has just one element.
 - Fix bug in the url when search something in the product page and click enter.
 - Footer component padding